evtloop: close the file descritors in d-tor, not after signaling shutdown
The problem here is that the OS may do not like it when file descriptors
are closed immediately after writing to them and relying on the fact
that the stuff written there should be read in another thread that holds
the same file descriptor. so, when shutting down the dispatcher we close
the pipe-fds into the d-tor, not right afterafter writing the exit
message.