evtloop: listen to input events on read-fd only
the fds in _fdunlock are pipe-fds, therefore one is for read and the
other one for writing. since we write on _fdunlock[1] we only have to
listen for input events on _fdunlock[0] to wake up on exit.
There is absolutely no need to listen to pri or out events since there
is only one event signalled with this fds - wake up for shutdown.